backend/requirements.txt
asgiref==3.3.1
astroid==2.4.2
babelfish==0.5.5
beautifulsoup4==4.9.1
bs4==0.0.1
click==7.1.2
decorator==4.4.2
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
dogpile.cache==1.1.1
enzyme==0.4.1
filelock==3.0.12
Flask==1.1.2
grip==4.5.2
guessit==3.1.1
isort==5.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
path-and-address==2.0.1
pbr==5.5.1
Pygments==2.6.1
pylint==2.6.0
pysrt==1.1.2
python-dateutil==2.8.1
pytz==2020.4
rarfile==4.0
rebulk==2.0.1
soupsieve==2.0.1
sqlparse==0.4.1
stevedore==3.2.2
subliminal==2.1.0
transliterate==1.10.2
virtualenv==20.0.14
Werkzeug==1.0.1
wrapt==1.12.1
